Crate riddle_platform_common

Riddle crate containing common utilities and types needed by platform implementations and which other crates can use to interact with a platform service without needing to know or genericize for the platform system type being used.

Most types in here are either consumed or reexported through the concrete platform crate (riddle_platform_winit). There should be very few cases where this crate will need to be depended on directly.

Modules

doctest
DO NOT RELY ON THIS MODULE

traits
Traits needed in order to implement a riddle compatible platform system.

Structs

LogicalPosition
A 2d position in logical screen units.

LogicalSize
A 2d size in logical screen units.

LogicalVec2
A 2d vector in logical screen units.

WindowId
A platform system independent window identifier.

Enums

MouseButton
Mouse button identifiers.

PlatformEvent

Scancode
Supported scancodes.

VirtualKey
Supported virtual keys.