Enum riddle_platform_common::VirtualKey

#[repr(u16)]
pub enum VirtualKey {

    Unknown,
    Escape,
    One,
    Two,
    Three,
    Four,
    Five,
    Six,
    Seven,
    Eight,
    Nine,
    Zero,
    Minus,
    Equal,
    Backspace,
    Tab,
    Q,
    W,
    E,
    R,
    T,
    Y,
    U,
    I,
    O,
    P,
    LeftBrace,
    RightBrace,
    Enter,
    LeftControl,
    A,
    D,
    S,
    F,
    G,
    H,
    J,
    K,
    L,
    Semicolon,
    Apostrope,
    Grave,
    LeftShift,
    Backslash,
    Z,
    X,
    C,
    V,
    B,
    N,
    M,
    Comma,
    Dot,
    Slash,
    RightShift,
    KeyPadAsterick,
    LeftAlt,
    Space,
    CapsLock,
    F1,
    F2,
    F3,
    F4,
    F5,
    F6,
    F7,
    F8,
    F9,
    F10,
    NumLock,
    ScrollLock,
    KeyPad7,
    KeyPad8,
    KeyPad9,
    KeyPadMinus,
    KeyPad4,
    KeyPad5,
    KeyPad6,
    KeyPadPlus,
    KeyPad1,
    KeyPad2,
    KeyPad3,
    KeyPad0,
    KeyPadDot,
    NonUSBackslash,
    F11,
    F12,
    KeyPadEnter,
    RightAlt,
    Home,
    Up,
    PageUp,
    Left,
    Right,
    End,
    Down,
    PageDown,
    Insert,
    Delete,
    KeyPadEqual,
    KeyPadPlusMinus,

}

Supported virtual keys.

Virtual keys represent the intended meaning of the key, and have no relation to where the key physically is on the keyboard. Use virtual key where the meaning of the key is most important (textual input). When meaning matters less, but physical location is more important (WASD-like control schemes) use crate::Scancode.

Variants

Unknown

Escape

One

Two

Three

Four

Five

Six

Seven

Eight

Nine

Zero

Minus

Equal

Backspace

Tab

Q

W

E

R

T

Y

U

I

O

P

LeftBrace

RightBrace

Enter

LeftControl

A

D

S

F

G

H

J

K

L

Semicolon

Apostrope

Grave

LeftShift

Backslash

Z

X

C

V

B

N

M

Comma

Dot

Slash

RightShift

KeyPadAsterick

LeftAlt

Space

CapsLock

F1

F2

F3

F4

F5

F6

F7

F8

F9

F10

NumLock

ScrollLock

KeyPad7

KeyPad8

KeyPad9

KeyPadMinus

KeyPad4

KeyPad5

KeyPad6

KeyPadPlus

KeyPad1

KeyPad2

KeyPad3

KeyPad0

KeyPadDot

NonUSBackslash

F11

F12

KeyPadEnter

RightAlt

Home

Up

PageUp

Left

Right

End

Down

PageDown

Insert

Delete

KeyPadEqual

KeyPadPlusMinus